+++ tests/Plugin/BufferedDelete/BufferedDeleteTest.php
@@ -7,6 +7,7 @@ use Solarium\Plugin\BufferedDelete\Delete\Id as DeleteById;
 use Solarium\Plugin\BufferedDelete\Delete\Query as DeleteQuery;
 use Solarium\Plugin\BufferedDelete\Event\AddDeleteById;
 use Solarium\Plugin\BufferedDelete\Event\AddDeleteQuery;
+use Solarium\Tests\Integration\TestClientFactory;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
 class BufferedDeleteTest extends BufferedDeleteLiteTest
@@ -20,6 +21,14 @@ class BufferedDeleteTest extends BufferedDeleteLiteTest
      * @var BufferedDelete
      */
     protected $plugin;
+
+    public function testInitPlugin()
+    {
+        $client = TestClientFactory::createWithCurlAdapter();
+        $plugin = $client->getPlugin('buffereddelete');
+
+        $this->assertInstanceOf(BufferedDelete::class, $plugin);
+    }
 
     public function testAddDeleteByIdEventIsTriggered()
     {